Add configurable pin mapping, all-SPI expanders, custom I/O, sensor descriptors, and Speeduino features
Squashed from flexi-pin branch (8 commits):

- Migrate all I/O expanders from 4x I2C MCP23017 to 6x SPI MCP23S17
  with shared CS and HAEN hardware addressing on HSPI bus at 10MHz
- Add 19 configurable GPIO pins with board presets, duplicate detection,
  and bootstrap pin labels (GPIO 0, 45, 46)
- Add custom I/O pins with output rules (threshold, range, delta triggers)
- Add virtual sensor descriptors (engine state, output state sources)
  with dedicated /sensors page and example configs
- Add Speeduino-inspired features: overdwell protection, fuel pump control,
  ASE (after-start enrichment), DFCO (decel fuel cut), CLT-based rev limit,
  dynamic fault curves, tooth logging
- Move pin config and peripherals from config page to dedicated pins page
- Fix uint8_t truncation for SPI expander pins (264+), pinLabel() priority,
  stale I2C references in /pins endpoint and sensors help text

+++ include/CJ125Controller.h
@@ -19,7 +19,7 @@ public:
 
     CJ125Controller(SPIClass* spi);
 
-    void begin(uint8_t spiSS1, uint8_t spiSS2,
+    void begin(uint16_t spiSS1, uint16_t spiSS2,
                uint8_t heaterOut1, uint8_t heaterOut2,
                uint8_t uaPin1, uint8_t uaPin2);
     void setADS1115(ADS1115Reader* ads) { _ads = ads; }
@@ -74,7 +74,7 @@ private:
         uint16_t urValue = 0;      // Current UR ADC (10-bit equiv)
         uint16_t uaValue = 0;      // Current UA ADC (10-bit equiv)
         uint16_t diagStatus = 0;
-        uint8_t spiSS = 0;         // Chip select pin (MCP23017)
+        uint16_t spiSS = 0;        // Chip select pin (MCP23S17)
         uint8_t heaterPin = 0;     // LEDC PWM output
         uint8_t ledcChannel = 0;   // LEDC channel number
         uint8_t uaPin = 0;         // ADC input
